Added simulation scenarios
Added code to evaluate the methods in 2 new scenarios; No association and Misspecified as well as code to compile the results. More results were added as well as 2 new data sets. I also added an example csv file to the shiny app tutorial as well as  an output csv file from using the example.